+++ CHANGELOG.md
@@ -30,6 +30,7 @@
   * Change `DropInPaymentMethodType#getCardTypes()` parameter from `Set` to `List`
   * Replace `VaultManagerActivity` with `VaultManagerFragment`
   * Remove `clientToken` and `tokenizationKey` from `DropInRequest` in favor of instantiating `DropInClient` with authorization
+  * Remove deprecated `amount` parameter from `DropInRequest`
   * Replace `DropInActivity.EXTRA_ERROR` with `DropInResult.EXTRA_ERROR`
 
 **Note:** The credentials for integrating with 3DS have changed. If you are using 3DS please update the credentials in your app-level `build.gradle`